requirements.txt
PySide6>=6.6
Pillow>=10.0.0
PyInstaller>=6.6

